Almost finished
Don't ask. It's been a while since I have had time to work on this. Some of it is my fault, some of it is just life. I had surgery on my hand, went on holiday, and did a lot of learning about advanced use of excel (which is oddly resemblent of moden code style, but that is understandable).

I have managed to get the game working apart from the win screen, but I have been staring at it all day I will look at it in the morning. All the variables are linked so you could start messing around with the grid size if you could find a way of relaunching the browser window, but that takes more javascript knowledge than I currently have. As always I have added lots of rambling notes but I think I am going to drop the """ thing inside functions as it looks silly and confuses code skulptor.