Finished wk6
Made a 1D version of pairs with objects, the realised how simple it was to make the 2D version then made it. Objects to make life a lot easier, and I am starting to get my head around it. The rule is build the first version well and make elegant changes.