Finished blackjack
Finished version 1. I have a non pretty but fairly robust version with details of the assignment, the stuff I have learnt, an analysis of the instructions and testing method, and ideas on how to improve it.

There are probably a few more bits to neaten or tweak but I think most of it is there.