Tweaks and consolidation
Had a look through all the stuff I have learnt in the course, finding answers to any questions I have missed, summarising key points, and tweaking the """ as neither codeskultopr nor atom really seem to like them. Not been as regulr with commits as had bad news from the hospital and yeah it's just the time of year for not being focused. Must build good habbits!