Created cheat sheet
Got annoying with scrolling up  and down the page trying to find the exact name of the various variables and attributes so made a cheat sheet with all the important ones on. I love having multiple screens. It also helps mentally recombine all the parts that are abstracted in oop.