Almost finished
Got all the basics in place, just more tweaking as wanted and getting the explosion animation in. Still finding the notes sheet handy as it stops me losing my place when scroling through code looking up a name of some kind. Tomorrow is adding in the animation then I want to analyse the structure of the instructions they give you. It's all well and good ebing able to follow instructions but soon I am going to have to write my own.